requirements.txt
matrix-nio>=0.24.0
aiohttp>=3.9.0
asyncio
